refactor folder structure of master role module
