+++ src/controllers/roleController/index.js
@@ -1,0 +1,7 @@
+module.exports = {
+  findAll: require("./findAll"),
+  findByID: require("./findByID"),
+  create: require("./create"),
+  update: require("./update"),
+  delete: require("./delete"),
+};
