+++ src/routes/roleRouter.js
@@ -1,18 +1,12 @@
 const router = require("express").Router();
-const {
-  getAllRoles,
-  getOneRole,
-  updateRole,
-  deleteRole,
-  createRole,
-} = require("../controllers/mstRoleController");
+const roleController = require("../controllers/roleController");
 const { superadminAuth } = require("../pkg/middlewares/auth");
 
-router.get("/roles", superadminAuth, getAllRoles); // get all roles
-router.post("/roles/:id", superadminAuth, createRole); // get one role
-router.get("/roles/:id", superadminAuth, getOneRole); // get one role
-router.patch("/roles/:id", superadminAuth, updateRole); // get one role
-router.delete("/roles/:id", superadminAuth, deleteRole); // get one role
+router.get("/roles", superadminAuth, roleController.findAll); // find all roles
+router.post("/roles/:id", superadminAuth, roleController.create); // create new role
+router.get("/roles/:id", superadminAuth, roleController.findByID); // find one role by id
+router.patch("/roles/:id", superadminAuth, roleController.update); // get one role
+router.delete("/roles/:id", superadminAuth, roleController.delete); // get one role
 
 // example documentation with swagger-jsdoc
 /**
